[FLINK-8029] Create WebMonitorEndpoint
The WebMonitorEndpoint is the common rest endpoint used for serving
the web frontend REST calls. It will be used by the Dispatcher and
the JobMaster to fuel the web frontend.

This closes #4987.
